requirements.txt
pyvts
sherpa-onnx
sounddevice
pyyaml
loguru
requests
tqdm
onnxruntime
